move analytic gaussian to validator
conversion to accuracy is handled, conversion to epsilon is still up in the air
